Converter is implemented(radius<->force, size<->size_id). some bug fixes, improvement in effects

+++ src/game/effects/ShockWaveEffect.hpp
@@ -47,6 +47,6 @@ class ShockWaveEffect
       	
 };
 
-ShockWaveEffect* getNewShockWave(int, bool dynamic = true);          
+ShockWaveEffect* getNewShockWave(float, bool dynamic = true);          
     
 #endif
